Revert weight_info
