--- xpallets/assets-registrar/src/default_weights.rs
@@ -1,28 +1,0 @@
-// Copyright 2019-2020 ChainX Project Authors. Licensed under GPL-3.0.
-
-use frame_support::weights::{constants::RocksDbWeight as DbWeight, Weight};
-
-use crate::WeightInfo;
-
-impl WeightInfo for () {
-    fn register() -> Weight {
-        (331853000 as Weight)
-            .saturating_add(DbWeight::get().reads(7 as Weight))
-            .saturating_add(DbWeight::get().writes(8 as Weight))
-    }
-    fn deregister() -> Weight {
-        (177795000 as Weight)
-            .saturating_add(DbWeight::get().reads(6 as Weight))
-            .saturating_add(DbWeight::get().writes(4 as Weight))
-    }
-    fn recover() -> Weight {
-        (238223000 as Weight)
-            .saturating_add(DbWeight::get().reads(7 as Weight))
-            .saturating_add(DbWeight::get().writes(5 as Weight))
-    }
-    fn update_asset_info() -> Weight {
-        (93039000 as Weight)
-            .saturating_add(DbWeight::get().reads(1 as Weight))
-            .saturating_add(DbWeight::get().writes(1 as Weight))
-    }
-}
